requirements.txt
tf-shell==0.1.33
dp-accounting==0.4.4
nltk==3.9.1
tensorboard-plugin-profile
# extra tensorboard-plugin-profile deps
etils
importlib-resources
# IMDB dataset
tensorflow_hub
tensorflow_datasets
keras_tuner